+++ src/storage/page_manager/page_manager_iter.c
@@ -28,7 +28,7 @@ bool page_iterator_has_next(PageIterator *self) {
     ASSERT_ARG_NOT_NULL(self)
 
     uint32_t pages_count = page_manager_get_pages_count(self->page_manager);
-    return self->next_page_id.id < pages_count;
+    return self->next_page_id.id < (int32_t) pages_count;
 }
 
 Result page_iterator_next(PageIterator *self, Page **result) {
